编译通过

+++ source/utils/utils.hpp
@@ -13,7 +13,7 @@ bool oglError();
 
 template <typename... Args>
 inline std::string string_format(const std::string& format, Args... args) {
-	size_t size = std::snprintf(nullptr, 0, format.c_str(), args...) + 1;
+	size_t size = (size_t)std::snprintf(nullptr, 0, format.c_str(), args...) + (size_t)1;
 	std::unique_ptr<char[]> buf(new char[size]);
 	std::snprintf(buf.get(), size, format.c_str(), args...);
 	return std::string(buf.get(), buf.get() + size - 1);
